feat: validate uuid on delete student repository

+++ src/infra/db/sequelize/student/repository/delete-student.repository.ts
@@ -1,10 +1,13 @@
 import { BadRequestException } from "@/domain/@shared/exceptions";
 import { DeleteRepository } from "@/domain/@shared/repository/repository";
 import Messages from "@/domain/@shared/util/messages";
+import { validate } from "uuid";
 import { StudentModel } from "../model";
-
 export class SequelizeDeleteStudentRepository implements DeleteRepository {
   async delete(id: string): Promise<void> {
+    if (!validate(id)) {
+      throw new BadRequestException(Messages.INVALID_ID);
+    }
     try {
       await StudentModel.destroy({
         where: { id },
